requirements.txt
flask>=2.3,<3
rembg>=2.0.67
Pillow>=10.0
gunicorn>=21.2
onnxruntime>=1.15.1
